requirements.txt
streamlit
openai
langchain
langchain-community
requests
python-dotenv